Feature IDENTIFIER expression [parser]
This commit sets up several types to scaffold the
structure of the Pratt parser. It firstly defines as
an enum the precedence values bound to operators.

Secondly, two types of function have been defined:

    - PrefixParserFunction
    - InfixParserFunction

The former will be in charge of parsing operators that
are in front of the next expression E.g: !<expression>,
while the latter will do it for operators sitting on the
middle of expressions, such as <expressionA> + <expressionB>.

These types of functions shall be linked to tokens so as to
associate token literals (operators) to parsing functions. This
association is made via the respective maps added to the parser
as properties. Parser functions are registered during parser
creation/instantiation.

On the other hand, a wrapper for expressions that produce values
when statements are expected (let's say, as direct childs of
ProgramNode or BlockNode) has been added: ExpressionStatement. It may sound
conflicting doing so, but this way ensures that these expressions
can be accessed through the same interface that statements are expected.
Moreover, statements like "1 + 2;" or "!(1 > 2)" are perfectly valid,
as expected of an interpreted language.