Feature PREFIX expression parsing [parser]
This commit adds support for prefix expressions such as
-2 or !1. In order to achieve this, a new node type has
been added to our AST: PrefixExpression. This node will
have the usual token as well as a field for the prefix
operator (bang and minus so far) and a field to hold
the expression at the right of the operator.

As a side effect this commit isolates testing integer
literals into a separate and private function local to
the parser testing module in the parser.