Feature INFIX expression parsing [parser][ast]
This commit completes the remaining half pending
duty of the parser: parsing infix expressions. These
expressions are those where the operator is only meaningful
with two operands.

For that reason, its AST node counts with 4 fields:

- The typical token
- Left: The expression at the left side of the operator
- Operator: The string representation of the operator
- Right: The expression at the right side of the operator

Since a language needs formal proof of operator evaluation
order, the implementation needed to parse left denotations
requires to be aware of operator precedences. Those are
defined as incremental constants in a enum-like form. Later
on, token types are associated each precedence.

To summarize, when parsing inf. expressions, recursive calls
are made while the precedence value of the current token under
examination is lower than the precedence of the next one. When
this happens, we retrieve the parsing function associated with
the next token. This function is passed the current node to build
a wrapper node where it will take the left position. This process
repeats itself until an operator with higher precedence is found.
In this case, the current node is returned. This outermost infix
node will be finally wrapper within a ExpressionStatement node,
and the next statement will be parsed, creating another branch
hanging from either a ProgramNode or a BlockStatement node. The
former is the root node while the later is still unimplemented.