Feature CALL expression [parser]
This commit adds a new AST node to represent function
invocations of the form:

```<expression>([<expresssion>...]) ```

This node is made up with a list of `Arguments` (which shall
be expressions) and a `Function` member at the left side which
should evaluate to a function object.

Calling expressions have the highest precedence in the ast, as it
has been reflected with the new value added in the precedences list
and the new test cases added to operator precedence parsing test
suite.

Calling expressions adheres to infix operands/operators location so
a the brand new "parseCallExpression(<left Expression>) has been
added to the parse infix functions registry. This function is given
as argument the expression at the left side of the LPAREN token, so
it is only responsible to parse subsequent function arguments.