Feature INTEGER objects [eval]
This commit scaffolds the code base to evaluate
AST nodes. Integer ones, in particular. But in order
to have green tests, ast.Program and ast.ExpressionStatement
had to be implemented as they are included the minimal
AST tree structure.

The REPL finally deserves the E!!... And that has been
reflected in the repl.go file where the final evaluated
object is inspected, displaying its value to the out
buffer.