Feature FUNCTION/CALL expressions [eval]
This commit implements both function and call
expressions evaluation.

In order to provide a clean 'closure' implementation, the
environment that functions hold at the moment of their
definition is extended by providing a pointer to the outer
environment. That is the reason why function objects carry
their own pointer to an environment; If using the current
env when being called (when invoking the calling expression)
they may never have access to that scope again.

There are some pending tasks:

    - Extend the test suite: Error handling and Return values
      - Referencing wrong params
      - Referencing identifiers that never existed...
    - Outer environment lookup