Feture LENGTH (len) builtin function
Builtin functions are those that are available
with the language as is, meaning that it is impossible
for the user to define them since they are direct
bindings with the host language and therefore more
powerfull.

This commit adds the first one of many: _len_. This will
return the element count for list-like or iterable objects.
Additionally, support for the "string" datatype has been
implemented too.

Since builtins are mapped to identifiers, the only meaningful
change was to extend de 'evalIdentifierExpression' so as to
look into builtins prior to raise a reference error if not
binding is found in the environment. On the other hand, applyFunction
has been extendend to not only _Function_ objects, but
_Builtin_ ones too.