Feature HASH LITERAL expressions [parser]
This commit fully adds support for hash literal
expressions. Keys can be any expressions, as well
as values are (that is to be restricted on eval time
to hashable objects only)

Also, for `StringLiteral` stringification method has
be altered so as to remove useless enclosing quotes that
did more harm than good.